aws/lambda/cross_repo_ci_relay/requirements.txt
fastapi
uvicorn
PyJWT[cryptography]
requests
PyYAML
PyGithub
clickhouse-connect
python-dotenv
